+++ crates/sncast/tests/e2e/account/create.rs
@@ -589,15 +589,12 @@ pub async fn test_happy_case_keystore_hex_format() {
     assert!(contents.contains("\"legacy\": false"));
 }
 
-#[test_case("oz"; "oz_account_type")]
-#[test_case("argent"; "argent_account_type")]
-#[test_case("braavos"; "braavos_account_type")]
 #[tokio::test]
-pub async fn test_happy_case_default_name_generation(account_type: &str) {
-    let temp_dir = tempdir().expect("Unable to create a temporary directory");
+pub async fn test_happy_case_default_name_generation() {
+    let tempdir = tempdir().expect("Unable to create a temporary directory");
     let accounts_file = "accounts.json";
 
-    let args = vec![
+    let create_args = vec![
         "--accounts-file",
         accounts_file,
         "account",
@@ -606,16 +603,22 @@ pub async fn test_happy_case_default_name_generation(account_type: &str) {
         URL,
         "--salt",
         "0x1",
-        "--type",
-        account_type,
     ];
 
-    let snapbox = runner(&args).current_dir(temp_dir.path());
-    let output = snapbox.assert().success();
+    let delete_args = vec![
+        "--accounts-file",
+        &accounts_file,
+        "account",
+        "delete",
+        "--name",
+        "account-2",
+        "--network",
+        "alpha-sepolia",
+    ];
 
-    assert_stdout_contains(
-        output,
-        indoc! {r"
+    for i in 0..3 {
+        let snapbox = runner(&create_args).current_dir(tempdir.path());
+        snapbox.assert().stdout_matches(formatdoc! {r"
         command: account create
         add_profile: --add-profile flag was not set. No profile added to snfoundry.toml
         address: 0x0[..]
@@ -623,15 +626,50 @@ pub async fn test_happy_case_default_name_generation(account_type: &str) {
         message: Account successfully created. Prefund generated address with at least <max_fee> STRK tokens or an equivalent amount of ETH tokens. It is good to send more in the case of higher demand.
 
         After prefunding the address, run:
-        sncast --accounts-file accounts.json account deploy --url http://127.0.0.1:5055/rpc --name account-1 --fee-token strk
+        sncast --accounts-file accounts.json account deploy --url http://127.0.0.1:5055/rpc --name account-{id} --fee-token strk
 
         To see account creation details, visit:
         account: [..]
-        "},
-    );
+    ", id = i + 1});
+    }
 
-    let contents = fs::read_to_string(temp_dir.path().join(accounts_file))
+    let contents = fs::read_to_string(tempdir.path().join(accounts_file))
         .expect("Unable to read created file");
+
+    assert!(contents.contains("account-1"));
+    assert!(contents.contains("account-2"));
+    assert!(contents.contains("account-3"));
+
+    let snapbox = runner(&delete_args).current_dir(tempdir.path()).stdin("Y");
+    snapbox.assert().success().stdout_matches(indoc! {r"
+        command: account delete
+        result: Account successfully removed
+    "});
+
+    let contents_after_delete = fs::read_to_string(tempdir.path().join(accounts_file))
+        .expect("Unable to read created file");
+
+    assert!(!contents_after_delete.contains("account-2"));
+
+    let snapbox = runner(&create_args).current_dir(tempdir.path());
+    snapbox.assert().stdout_matches(indoc! {r"
+        command: account create
+        add_profile: --add-profile flag was not set. No profile added to snfoundry.toml
+        address: 0x0[..]
+        max_fee: [..]
+        message: Account successfully created. Prefund generated address with at least <max_fee> STRK tokens or an equivalent amount of ETH tokens. It is good to send more in the case of higher demand.
+
+        After prefunding the address, run:
+        sncast --accounts-file accounts.json account deploy --url http://127.0.0.1:5055/rpc --name account-2 --fee-token strk
+
+        To see account creation details, visit:
+        account: [..]
+    "});
+
+    let contents = fs::read_to_string(tempdir.path().join(accounts_file))
+        .expect("Unable to read created file");
+
+    assert!(contents.contains("account-2"));
 
     let expected = json!(
         {
@@ -644,8 +682,28 @@ pub async fn test_happy_case_default_name_generation(account_type: &str) {
                     "private_key": "0x[..]",
                     "public_key": "0x[..]",
                     "salt": "0x1",
-                    "type": get_formatted_account_type(account_type)
-                }
+                    "type": "open_zeppelin"
+                },
+                "account-2": {
+                    "address": "0x[..]",
+                    "class_hash": "0x[..]",
+                    "deployed": false,
+                    "legacy": false,
+                    "private_key": "0x[..]",
+                    "public_key": "0x[..]",
+                    "salt": "0x1",
+                    "type": "open_zeppelin"
+                },
+                "account-3": {
+                    "address": "0x[..]",
+                    "class_hash": "0x[..]",
+                    "deployed": false,
+                    "legacy": false,
+                    "private_key": "0x[..]",
+                    "public_key": "0x[..]",
+                    "salt": "0x1",
+                    "type": "open_zeppelin"
+                },
             }
         }
     );